test: 服务端超时抛出
