+++ src/flow/server.py
@@ -1,34 +1,10 @@
 from asyncio import StreamWriter
 import asyncio
 import json
-from typing import Any, Callable, Dict, Tuple, Union
-from ..exception.tcp import ServiceExistException, ServiceNotFoundException
+from typing import Any, Callable, Union
+from ..tool.server_tool import ServerRegister
 from ..tool.loop_tool import OrderApi
 from ..tool.bytes_tool import CODING
-
-
-class ServerRegister:
-    __TABLE: Dict[str, Tuple[Callable, bool]] = dict()
-
-    def __init__(self, path: Union[str, None] = None):
-        self.__path = path
-        pass
-
-    def __call__(self, func: Callable):
-        path = f'{ "" if self.__path is None else self.__path}/{func.__name__}'
-        if self.__class__.__TABLE.get(path, None) is not None:
-            raise ServiceExistException(f'服务已存在，无法注册:{path}')
-        self.__class__.__TABLE[path] = func, asyncio.iscoroutinefunction(func)
-        return func
-
-    @classmethod
-    async def call(cls, path, *args, **kwds) -> Tuple[Callable, bool]:
-        if cls.__TABLE.get(path, None) is None:
-            raise ServiceNotFoundException(f'服务不存在:{path}')
-        func, is_async = cls.__TABLE[path]
-        func_res = func(*args, **kwds)
-        return await func_res if is_async else func_res
-    pass
 
 
 class FlowSendServer(OrderApi):
@@ -59,6 +35,7 @@ class ServiceMapping:
         try:
             return await asyncio.wait_for(ServerRegister.call(service_name, *args, **kwds), 1)
         except asyncio.TimeoutError:
+            # TODO: 记录所有超时
             raise Exception(f'调用服务超时:{id}|{service_name}|{args}|{kwds}')
     pass
