+++ src/tool/server_tool.py
@@ -1,0 +1,28 @@
+import asyncio
+from typing import Callable, Dict, Tuple, Union
+
+from ..exception.tcp import ServiceExistException, ServiceNotFoundException
+
+
+class ServerRegister:
+    __TABLE: Dict[str, Tuple[Callable, bool]] = dict()
+
+    def __init__(self, path: Union[str, None] = None):
+        self.__path = path
+        pass
+
+    def __call__(self, func: Callable):
+        path = f'{ "" if self.__path is None else self.__path}/{func.__name__}'
+        if self.__class__.__TABLE.get(path, None) is not None:  # pragma: no cover
+            raise ServiceExistException(f'服务已存在，无法注册:{path}')
+        self.__class__.__TABLE[path] = func, asyncio.iscoroutinefunction(func)
+        return func
+
+    @classmethod
+    async def call(cls, path, *args, **kwds) -> Tuple[Callable, bool]:
+        if cls.__TABLE.get(path, None) is None:
+            raise ServiceNotFoundException(f'服务不存在:{path}')
+        func, is_async = cls.__TABLE[path]
+        func_res = func(*args, **kwds)
+        return await func_res if is_async else func_res
+    pass
